Move internal opts to a package
