+++ lib/scala/cli/src/main/scala/org/enso/cli/internal/opts/OptionalPositionalArgument.scala
@@ -1,5 +1,6 @@
-package org.enso.cli.internal
+package org.enso.cli.internal.opts
 
+import org.enso.cli.internal.ParserContinuation
 import org.enso.cli.{Argument, OptsParseError}
 
 class OptionalPositionalArgument[A: Argument](
