+++ lib/scala/cli/src/main/scala/org/enso/cli/internal/opts/PositionalArgument.scala
@@ -1,5 +1,6 @@
-package org.enso.cli.internal
+package org.enso.cli.internal.opts
 
+import org.enso.cli.internal.ParserContinuation
 import org.enso.cli.{Argument, OptsParseError}
 
 class PositionalArgument[A: Argument](
